+++ pkg/applicationserver/applicationserver_test.go
@@ -862,7 +862,7 @@ hardware_versions:
 				defer wg.Done()
 				err := ptc.Connect(ctx, t, registeredApplicationID, registeredApplicationKey, chs)
 				if !errors.IsCanceled(err) {
-					t.Fatalf("Expected context canceled, but have %v", err)
+					t.Errorf("Expected context canceled, but have %v", err)
 				}
 			}()
 			// Wait for connection to establish.
@@ -1906,6 +1906,401 @@ hardware_versions:
 
 			cancel()
 			wg.Wait()
+		})
+	}
+}
+
+func TestSkipPayloadCrypto(t *testing.T) {
+	a := assertions.New(t)
+
+	// This application will be added to the Entity Registry and to the link registry of the Application Server so that it
+	// links automatically on start to the Network Server.
+	registeredApplicationID := ttnpb.ApplicationIdentifiers{ApplicationID: "foo-app"}
+	registeredApplicationKey := "secret"
+
+	// This device gets registered in the device registry of the Application Server.
+	registeredDevice := &ttnpb.EndDevice{
+		EndDeviceIdentifiers: ttnpb.EndDeviceIdentifiers{
+			ApplicationIdentifiers: registeredApplicationID,
+			DeviceID:               "foo-device",
+			JoinEUI:                eui64Ptr(types.EUI64{0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff}),
+			DevEUI:                 eui64Ptr(types.EUI64{0x42, 0x42, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff}),
+		},
+	}
+
+	ctx := test.Context()
+	is, isAddr := startMockIS(ctx)
+	js, jsAddr := startMockJS(ctx)
+	ns, nsAddr := startMockNS(ctx, func(md rpcmetadata.MD) bool {
+		return md.AuthType == "Bearer" && md.AuthValue == registeredApplicationKey
+	})
+
+	// Register the application in the Entity Registry.
+	is.add(ctx, registeredApplicationID, registeredApplicationKey)
+
+	// Register some sessions in the Join Server. Sometimes the keys are sent by the Network Server as part of the
+	// join-accept, and sometimes they are not sent by the Network Server so the Application Server gets them from the
+	// Join Server.
+	js.add(ctx, *registeredDevice.DevEUI, []byte{0x11}, ttnpb.KeyEnvelope{
+		// AppSKey is []byte{0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11}.
+		EncryptedKey: []byte{0xa8, 0x11, 0x8f, 0x80, 0x2e, 0xbf, 0x8, 0xdc, 0x62, 0x37, 0xc3, 0x4, 0x63, 0xa2, 0xfa, 0xcb, 0xf8, 0x87, 0xaa, 0x31, 0x90, 0x23, 0x85, 0xc1},
+		KEKLabel:     "test",
+	})
+	js.add(ctx, *registeredDevice.DevEUI, []byte{0x22}, ttnpb.KeyEnvelope{
+		// AppSKey is []byte{0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22}
+		EncryptedKey: []byte{0x39, 0x11, 0x40, 0x98, 0xa1, 0x5d, 0x6f, 0x92, 0xd7, 0xf0, 0x13, 0x21, 0x5b, 0x5b, 0x41, 0xa8, 0x98, 0x2d, 0xac, 0x59, 0x34, 0x76, 0x36, 0x18},
+		KEKLabel:     "test",
+	})
+	js.add(ctx, *registeredDevice.DevEUI, []byte{0x33}, ttnpb.KeyEnvelope{
+		// AppSKey is []byte{0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33, 0x33}
+		EncryptedKey: []byte{0x5, 0x81, 0xe1, 0x15, 0x8a, 0xc3, 0x13, 0x68, 0x5e, 0x8d, 0x15, 0xc0, 0x11, 0x92, 0x14, 0x49, 0x9f, 0xa0, 0xc6, 0xf1, 0xdb, 0x95, 0xff, 0xbd},
+		KEKLabel:     "test",
+	})
+	js.add(ctx, *registeredDevice.DevEUI, []byte{0x44}, ttnpb.KeyEnvelope{
+		// AppSKey is []byte{0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44, 0x44}
+		EncryptedKey: []byte{0x30, 0xcf, 0x47, 0x91, 0x11, 0x64, 0x53, 0x3f, 0xc3, 0xd5, 0xd8, 0x56, 0x5b, 0x71, 0xcb, 0xe7, 0x6d, 0x14, 0x2b, 0x2c, 0xf2, 0xc2, 0xd7, 0x7b},
+		KEKLabel:     "test",
+	})
+	js.add(ctx, *registeredDevice.DevEUI, []byte{0x55}, ttnpb.KeyEnvelope{
+		// AppSKey is []byte{0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55, 0x55}
+		EncryptedKey: []byte{0x56, 0x15, 0xaa, 0x22, 0xb7, 0x5f, 0xc, 0x24, 0x79, 0x6, 0x84, 0x68, 0x89, 0x0, 0xa6, 0x16, 0x4a, 0x9c, 0xef, 0xdb, 0xbf, 0x61, 0x6f, 0x0},
+		KEKLabel:     "test",
+	})
+
+	devsRedisClient, devsFlush := test.NewRedis(t, "applicationserver_test", "devices")
+	defer devsFlush()
+	defer devsRedisClient.Close()
+	deviceRegistry := &redis.DeviceRegistry{Redis: devsRedisClient}
+
+	linksRedisClient, linksFlush := test.NewRedis(t, "applicationserver_test", "links")
+	defer linksFlush()
+	defer linksRedisClient.Close()
+	linkRegistry := &redis.LinkRegistry{Redis: linksRedisClient}
+	_, err := linkRegistry.Set(ctx, registeredApplicationID, nil, func(_ *ttnpb.ApplicationLink) (*ttnpb.ApplicationLink, []string, error) {
+		return &ttnpb.ApplicationLink{
+			APIKey:            registeredApplicationKey,
+			SkipPayloadCrypto: &pbtypes.BoolValue{Value: true},
+		}, []string{"api_key", "skip_payload_crypto"}, nil
+	})
+	if err != nil {
+		t.Fatalf("Failed to set link in registry: %s", err)
+	}
+
+	c := componenttest.NewComponent(t, &component.Config{
+		ServiceBase: config.ServiceBase{
+			GRPC: config.GRPC{
+				Listen:                      ":0",
+				AllowInsecureForCredentials: true,
+			},
+			Cluster: cluster.Config{
+				IdentityServer: isAddr,
+				JoinServer:     jsAddr,
+				NetworkServer:  nsAddr,
+			},
+			KeyVault: config.KeyVault{
+				Provider: "static",
+				Static: map[string][]byte{
+					"test": {0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F},
+				},
+			},
+		},
+	})
+	config := &applicationserver.Config{
+		LinkMode: "all",
+		Devices:  deviceRegistry,
+		Links:    linkRegistry,
+	}
+	as, err := applicationserver.New(c, config)
+	if !a.So(err, should.BeNil) {
+		t.FailNow()
+	}
+
+	roles := as.Roles()
+	a.So(len(roles), should.Equal, 1)
+	a.So(roles[0], should.Equal, ttnpb.ClusterRole_APPLICATION_SERVER)
+
+	componenttest.StartComponent(t, c)
+	defer c.Close()
+	mustHavePeer(ctx, c, ttnpb.ClusterRole_NETWORK_SERVER)
+	mustHavePeer(ctx, c, ttnpb.ClusterRole_JOIN_SERVER)
+	mustHavePeer(ctx, c, ttnpb.ClusterRole_ENTITY_REGISTRY)
+
+	// Delay for the AS-NS link to establish.
+	time.Sleep(Timeout)
+
+	chs := &connChannels{
+		up:          make(chan *ttnpb.ApplicationUp, 1),
+		downPush:    make(chan *ttnpb.DownlinkQueueRequest),
+		downReplace: make(chan *ttnpb.DownlinkQueueRequest),
+		downErr:     make(chan error, 1),
+	}
+	creds := grpc.PerRPCCredentials(rpcmetadata.MD{
+		AuthType:      "Bearer",
+		AuthValue:     registeredApplicationKey,
+		AllowInsecure: true,
+	})
+	client := ttnpb.NewAppAsClient(as.LoopbackConn())
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+	stream, err := client.Subscribe(ctx, &registeredApplicationID, creds)
+	if err != nil {
+		t.Fatalf("Failed to subscribe: %v", err)
+	}
+	// Read upstream.
+	go func() {
+		for {
+			msg, err := stream.Recv()
+			if err != nil {
+				return
+			}
+			chs.up <- msg
+		}
+	}()
+	// Write downstream.
+	go func() {
+		for {
+			var err error
+			select {
+			case <-ctx.Done():
+				return
+			case req := <-chs.downPush:
+				_, err = client.DownlinkQueuePush(ctx, req, creds)
+			case req := <-chs.downReplace:
+				_, err = client.DownlinkQueueReplace(ctx, req, creds)
+			}
+			chs.downErr <- err
+		}
+	}()
+
+	for _, override := range []*pbtypes.BoolValue{
+		nil,
+		{Value: true},
+		{Value: false},
+	} {
+		hasOverride := map[bool]string{true: "Overrides", false: "NoOverride"}[override != nil]
+		effectiveSkip := override != nil && override.Value || override == nil
+
+		t.Run(fmt.Sprintf("%v/%v", hasOverride, effectiveSkip), func(t *testing.T) {
+			t.Run("Uplink", func(t *testing.T) {
+				ns.reset()
+				devsFlush()
+				deviceRegistry.Set(ctx, registeredDevice.EndDeviceIdentifiers, nil, func(_ *ttnpb.EndDevice) (*ttnpb.EndDevice, []string, error) {
+					dev := *registeredDevice
+					dev.SkipPayloadCryptoOverride = override
+					return &dev, []string{
+						"ids",
+						"formatters",
+						"skip_payload_crypto_override",
+					}, nil
+				})
+
+				for _, step := range []struct {
+					Name         string
+					Message      *ttnpb.ApplicationUp
+					AssertUp     func(t *testing.T, up *ttnpb.ApplicationUp)
+					AssertDevice func(t *testing.T, dev *ttnpb.EndDevice)
+				}{
+					{
+						Name: "JoinAccept",
+						Message: &ttnpb.ApplicationUp{
+							EndDeviceIdentifiers: withDevAddr(registeredDevice.EndDeviceIdentifiers, types.DevAddr{0x22, 0x22, 0x22, 0x22}),
+							Up: &ttnpb.ApplicationUp_JoinAccept{
+								JoinAccept: &ttnpb.ApplicationJoinAccept{
+									SessionKeyID: []byte{0x22},
+									AppSKey: &ttnpb.KeyEnvelope{
+										// AppSKey is []byte{0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22, 0x22}
+										EncryptedKey: []byte{0x39, 0x11, 0x40, 0x98, 0xa1, 0x5d, 0x6f, 0x92, 0xd7, 0xf0, 0x13, 0x21, 0x5b, 0x5b, 0x41, 0xa8, 0x98, 0x2d, 0xac, 0x59, 0x34, 0x76, 0x36, 0x18},
+										KEKLabel:     "test",
+									},
+								},
+							},
+						},
+						AssertUp: func(t *testing.T, up *ttnpb.ApplicationUp) {
+							a := assertions.New(t)
+							a.So(up, should.Resemble, &ttnpb.ApplicationUp{
+								EndDeviceIdentifiers: withDevAddr(registeredDevice.EndDeviceIdentifiers, types.DevAddr{0x22, 0x22, 0x22, 0x22}),
+								Up: &ttnpb.ApplicationUp_JoinAccept{
+									JoinAccept: &ttnpb.ApplicationJoinAccept{
+										SessionKeyID: []byte{0x22},
+									},
+								},
+								CorrelationIDs: up.CorrelationIDs,
+								ReceivedAt:     up.ReceivedAt,
+							})
+						},
+						AssertDevice: func(t *testing.T, dev *ttnpb.EndDevice) {
+							a := assertions.New(t)
+							a.So(dev.Session, should.BeNil)
+							a.So(dev.PendingSession, should.Resemble, &ttnpb.Session{
+								DevAddr: types.DevAddr{0x22, 0x22, 0x22, 0x22},
+								SessionKeys: ttnpb.SessionKeys{
+									SessionKeyID: []byte{0x22},
+									AppSKey: &ttnpb.KeyEnvelope{
+										EncryptedKey: []byte{0x39, 0x11, 0x40, 0x98, 0xa1, 0x5d, 0x6f, 0x92, 0xd7, 0xf0, 0x13, 0x21, 0x5b, 0x5b, 0x41, 0xa8, 0x98, 0x2d, 0xac, 0x59, 0x34, 0x76, 0x36, 0x18},
+										KEKLabel:     "test",
+									},
+								},
+								LastAFCntDown: 0,
+								StartedAt:     dev.PendingSession.StartedAt,
+							})
+						},
+					},
+					{
+						Name: "UplinkMessage/PendingSession",
+						Message: &ttnpb.ApplicationUp{
+							EndDeviceIdentifiers: withDevAddr(registeredDevice.EndDeviceIdentifiers, types.DevAddr{0x22, 0x22, 0x22, 0x22}),
+							Up: &ttnpb.ApplicationUp_UplinkMessage{
+								UplinkMessage: &ttnpb.ApplicationUplink{
+									SessionKeyID: []byte{0x22},
+									FPort:        22,
+									FCnt:         22,
+									FRMPayload:   []byte{0x01},
+								},
+							},
+						},
+						AssertUp: func(t *testing.T, up *ttnpb.ApplicationUp) {
+							a := assertions.New(t)
+							if effectiveSkip {
+								a.So(up, should.Resemble, &ttnpb.ApplicationUp{
+									EndDeviceIdentifiers: withDevAddr(registeredDevice.EndDeviceIdentifiers, types.DevAddr{0x22, 0x22, 0x22, 0x22}),
+									Up: &ttnpb.ApplicationUp_UplinkMessage{
+										UplinkMessage: &ttnpb.ApplicationUplink{
+											SessionKeyID: []byte{0x22},
+											FPort:        22,
+											FCnt:         22,
+											FRMPayload:   []byte{0x01},
+											AppSKey: &ttnpb.KeyEnvelope{
+												EncryptedKey: []byte{0x39, 0x11, 0x40, 0x98, 0xa1, 0x5d, 0x6f, 0x92, 0xd7, 0xf0, 0x13, 0x21, 0x5b, 0x5b, 0x41, 0xa8, 0x98, 0x2d, 0xac, 0x59, 0x34, 0x76, 0x36, 0x18},
+												KEKLabel:     "test",
+											},
+										},
+									},
+									CorrelationIDs: up.CorrelationIDs,
+									ReceivedAt:     up.ReceivedAt,
+								})
+							} else {
+								a.So(up, should.Resemble, &ttnpb.ApplicationUp{
+									EndDeviceIdentifiers: withDevAddr(registeredDevice.EndDeviceIdentifiers, types.DevAddr{0x22, 0x22, 0x22, 0x22}),
+									Up: &ttnpb.ApplicationUp_UplinkMessage{
+										UplinkMessage: &ttnpb.ApplicationUplink{
+											SessionKeyID: []byte{0x22},
+											FPort:        22,
+											FCnt:         22,
+											FRMPayload:   []byte{0xc1},
+										},
+									},
+									CorrelationIDs: up.CorrelationIDs,
+									ReceivedAt:     up.ReceivedAt,
+								})
+							}
+						},
+					},
+				} {
+					stepok := t.Run(step.Name, func(t *testing.T) {
+						ns.upCh <- step.Message
+						select {
+						case msg := <-chs.up:
+							if step.AssertUp != nil {
+								step.AssertUp(t, msg)
+							} else {
+								t.Fatalf("Expected no upstream message but got %v", msg)
+							}
+						case <-time.After(Timeout):
+							t.Fatal("Expected upstream timeout")
+						}
+						if step.AssertDevice != nil {
+							dev, err := deviceRegistry.Get(ctx, step.Message.EndDeviceIdentifiers, []string{"session", "pending_session"})
+							if !a.So(err, should.BeNil) {
+								t.FailNow()
+							}
+							step.AssertDevice(t, dev)
+						}
+					})
+					if !stepok {
+						t.FailNow()
+					}
+				}
+			})
+
+			t.Run("Downlink", func(t *testing.T) {
+				ns.reset()
+				devsFlush()
+				deviceRegistry.Set(ctx, registeredDevice.EndDeviceIdentifiers, nil, func(_ *ttnpb.EndDevice) (*ttnpb.EndDevice, []string, error) {
+					dev := *registeredDevice
+					dev.Session = &ttnpb.Session{
+						DevAddr: types.DevAddr{0x42, 0xff, 0xff, 0xff},
+						SessionKeys: ttnpb.SessionKeys{
+							SessionKeyID: []byte{0x11},
+							AppSKey: &ttnpb.KeyEnvelope{
+								EncryptedKey: []byte{0x1f, 0xa6, 0x8b, 0xa, 0x81, 0x12, 0xb4, 0x47, 0xae, 0xf3, 0x4b, 0xd8, 0xfb, 0x5a, 0x7b, 0x82, 0x9d, 0x3e, 0x86, 0x23, 0x71, 0xd2, 0xcf, 0xe5},
+								KEKLabel:     "test",
+							},
+						},
+					}
+					dev.SkipPayloadCryptoOverride = override
+					return &dev, []string{
+						"ids",
+						"session",
+						"skip_payload_crypto_override",
+					}, nil
+				})
+				t.Run("Push", func(t *testing.T) {
+					a := assertions.New(t)
+					for _, items := range [][]*ttnpb.ApplicationDownlink{
+						{
+							{
+								FPort:      11,
+								FRMPayload: []byte{0x1, 0x1, 0x1},
+							},
+							{
+								FPort:      22,
+								FRMPayload: []byte{0x2, 0x2, 0x2},
+							},
+						},
+					} {
+						chs.downPush <- &ttnpb.DownlinkQueueRequest{
+							EndDeviceIdentifiers: registeredDevice.EndDeviceIdentifiers,
+							Downlinks:            items,
+						}
+						time.Sleep(Timeout)
+						select {
+						case err := <-chs.downErr:
+							if !a.So(err, should.BeNil) {
+								t.FailNow()
+							}
+						default:
+							t.Fatal("Expected downlink error")
+						}
+						for i := 0; i < len(items); i++ {
+							select {
+							case up := <-chs.up:
+								a.So(up.Up, should.HaveSameTypeAs, &ttnpb.ApplicationUp_DownlinkQueued{})
+							default:
+								t.Fatalf("Expected upstream event")
+							}
+						}
+					}
+					res, err := as.DownlinkQueueList(ctx, registeredDevice.EndDeviceIdentifiers)
+					if a.So(err, should.BeNil) && a.So(res, should.HaveLength, 2) {
+						a.So(res, should.Resemble, []*ttnpb.ApplicationDownlink{
+							{
+								SessionKeyID:   []byte{0x11},
+								FPort:          11,
+								FCnt:           1,
+								FRMPayload:     []byte{0x1, 0x1, 0x1},
+								CorrelationIDs: res[0].CorrelationIDs,
+							},
+							{
+								SessionKeyID:   []byte{0x11},
+								FPort:          22,
+								FCnt:           2,
+								FRMPayload:     []byte{0x2, 0x2, 0x2},
+								CorrelationIDs: res[1].CorrelationIDs,
+							},
+						})
+					}
+				})
+			})
 		})
 	}
 }