+++ CHANGELOG.md
@@ -12,12 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setting custom gateway attributes in the console
 - Pub/Sub documentation.
 - Return informative well-known errors for standard network and context errors.
+- Skipping payload crypto on application-level via application link's `skip_payload_crypto` field.
 
 ### Changed
 
 - Event stream start and stop events are not published anymore.
 
 ### Deprecated
+
+- End device `skip_payload_crypto` field: it gets replaced by `skip_payload_crypto_override`.
 
 ### Removed